refactor(commands): Moves commands to new location
The `dadjoke` and `guidefetch` command packages have been relocated to
a `command` internal package.

This package is designed to contain future commands and structure them
by their package names in order to preserve their package names when
importing them.

This change aims to seperate internal bot service code, and its
commands.
